Run jekyll build with config variables
Jekyll related config variables are now passed to the `jeykyll
build` command.

Config vars are passed to the buildpack as an argument
(versus being set in the environment as it happens for commands
in the Procfile or when running one-off processes) as explained
in the link below (search for ENV_DIR):
https://devcenter.heroku.com/articles/buildpack-api

As a result the config variables I needed for jekyll build were not
being automatically passed to it.

Config variables are particularly useful whenever we have things
that we do not want to commit to a git repository (e.g. google analytics
keys) or whenever we want to load different assets in production/development
(e.g. minified CSS files). Examples:
https://github.com/marionzualo/www/blob/ec01d61035a2aaa832e985eee2651edafb4c2747/_includes/_head.html#L11
https://github.com/marionzualo/www/blob/ec01d61035a2aaa832e985eee2651edafb4c2747/_includes/_head.html#L17

Heroku docs on config vars:
https://devcenter.heroku.com/articles/config-vars